Rewrite more typescript
